+++ packages/apollo/src/commands/client/download-schema.ts
@@ -1,11 +1,12 @@
 import { flags } from "@oclif/command";
-import { introspectionFromSchema } from "graphql";
+import { introspectionFromSchema, printSchema } from "graphql";
 import { writeFileSync } from "fs";
 
 import { ClientCommand } from "../../Command";
 
 export default class SchemaDownload extends ClientCommand {
-  static description = "Download a schema from engine or a GraphQL endpoint.";
+  static description =
+    "Download a schema from engine or a GraphQL endpoint in JSON or SDL format";
 
   static flags = {
     ...ClientCommand.flags
@@ -14,7 +15,8 @@ export default class SchemaDownload extends ClientCommand {
   static args = [
     {
       name: "output",
-      description: "Path to write the introspection result to",
+      description:
+        "Path to write the introspection result to. Can be `.graphql`, `.gql`, `.graphqls`, or `.json`",
       required: true,
       default: "schema.json"
     }
@@ -23,17 +25,21 @@ export default class SchemaDownload extends ClientCommand {
   async run() {
     let result;
     let gitContext;
-    await this.runTasks(({ args, project, flags }) => [
-      {
-        title: `Saving schema to ${args.output}`,
-        task: async () => {
-          const schema = await project.resolveSchema({ tag: flags.tag });
-          writeFileSync(
-            args.output,
-            JSON.stringify(introspectionFromSchema(schema), null, 2)
-          );
+    await this.runTasks(({ args, project, flags }) => {
+      const extension = args.output.split(".").pop();
+      const isSDLFormat = ["graphql", "graphqls", "gql"].includes(extension);
+      return [
+        {
+          title: `Saving schema to ${args.output}`,
+          task: async () => {
+            const schema = await project.resolveSchema({ tag: flags.tag });
+            const formattedSchema = isSDLFormat
+              ? printSchema(schema)
+              : JSON.stringify(introspectionFromSchema(schema), null, 2);
+            writeFileSync(args.output, formattedSchema);
+          }
         }
-      }
-    ]);
+      ];
+    });
   }
 }